refactoring: use webpack + HMR unit tests + one build bundle per language
