+++ .gitignore
@@ -6,3 +6,4 @@ lib/
 .coveralls.yml
 bower_components/
 umd/
+output/
