--- .nycrc
@@ -1,33 +1,0 @@
-{
-  "check-coverage": true,
-  "per-file": true,
-  "lines": 100,
-  "statements": 100,
-  "functions": 100,
-  "branches": 100,
-  "include": [
-    "src/**/*.js"
-  ],
-  "exclude": [
-    "build/pegjs-parser.js",
-    "lib/**/*.js",
-    "test/**/*.spec.js"
-  ],
-  "ignore-class-method": "methodToIgnore",
-  "reporter": [
-    "lcov",
-    "text-summary",
-    "html"
-  ],
-  "require": [
-    "@babel/register"
-  ],
-  "extension": [
-    ".js"
-  ],
-  "sourceMap": true,
-  "instrument": true,
-  "cache": true,
-  "all": true,
-  "report-dir": "./coverage"
-}
