+++ package.json
@@ -5,14 +5,15 @@
   "main": "index.js",
   "types": "index.d.ts",
   "scripts": {
-    "build": "rm -rf build/* && mkdir -p build && node build.js && npm run compile",
-    "test": "NODE_ENV=test npm run lint && mocha --recursive --require @babel/register",
-    "compile": "babel src -d lib",
-    "browser": "browserify -p tinyify index.js -o umd/parser.min.js",
-    "lint": "npm run build && eslint src/",
-    "prepublishOnly": "npm run build && npm run browser",
-    "coverLocal": "nyc npm run test",
-    "cover": "nyc npm run test && nyc report --reporter=text-lcov | coveralls"
+    "start": "webpack --config webpack.config.js",
+    "build": "webpack --config webpack.config.js --prod",
+    "test": "mochapack test/**/*.spec.js",
+    "prepublishOnly": "npm run build",
+    "lint": "eslint src",
+    "coverLocal": "cross-env NODE_ENV=coverage nyc --reporter=lcov --reporter=text npm run test",
+    "cover:run": "cross-env NODE_ENV=coverage nyc --reporter=text-lcov npm run test",
+    "cover": "npm run cover:run && nyc report --reporter=text-lcov | coveralls",
+    "release": "npm publish output/prod"
   },
   "repository": {
     "type": "git",
@@ -50,25 +51,44 @@
   },
   "devDependencies": {
     "@babel/cli": "^7.5.5",
-    "@babel/core": "^7.5.5",
+    "@babel/core": "^7.10.5",
     "@babel/plugin-proposal-object-rest-spread": "^7.5.5",
-    "@babel/preset-env": "^7.5.5",
-    "@babel/register": "^7.5.5",
+    "@babel/preset-env": "^7.10.4",
+    "@babel/register": "^7.10.5",
     "@types/chai": "^4.1.7",
     "@types/mocha": "^8.0.0",
+    "babel-loader": "^8.1.0",
     "browserify": "^16.5.1",
     "chai": "^4.2.0",
+    "copy-webpack-plugin": "^6.0.3",
     "coveralls": "^3.0.13",
+    "cross-env": "^7.0.2",
     "eslint": "^6.8.0",
     "eslint-config-airbnb-base": "^14.0.0",
     "eslint-config-strict": "^14.0.1",
     "eslint-plugin-filenames": "^1.3.2",
     "eslint-plugin-import": "^2.16.0",
+    "istanbul-instrumenter-loader": "^3.0.1",
     "mocha": "^7.1.1",
+    "mochapack": "^2.0.3",
     "nyc": "15.1.0",
     "pegjs": "^0.10.0",
+    "pegjs-loader": "^0.5.6",
     "pre-commit": "^1.2.2",
-    "tinyify": "^2.5.2"
+    "rimraf": "^3.0.2",
+    "source-map-support": "^0.5.19",
+    "tinyify": "^2.5.2",
+    "vscode-mocha-hmr": "^1.0.0",
+    "webpack": "^4.43.0",
+    "webpack-cli": "^3.3.12",
+    "webpack-node-externals": "^1.7.2"
+  },
+  "nyc": {
+    "include": [
+      "src/**/*.js"
+    ],
+    "instrument": false,
+    "sourceMap": false
   },
   "dependencies": {
     "big-integer": "^1.6.48"
