+++ src/parser.single.js
@@ -1,0 +1,5 @@
+import {parse} from '../pegjs/mysql.pegjs';
+
+export default {
+    [PARSER_NAME]: parse,
+};
