+++ tests-index.js
@@ -1,0 +1,4 @@
+var req = require.context('./test', true, /\.spec$/);
+req.keys().forEach(req);
+
+require('vscode-mocha-hmr')(module);
